+++ src/components/ContentPage/ContentPage.js
@@ -1,10 +1,10 @@
 /*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
 
-import React, { PropTypes, Component } from 'react';
-import styles from './ContentPage.scss';
+import React, { Component, PropTypes } from 'react';
+import s from './ContentPage.scss';
 import withStyles from '../../decorators/withStyles';
 
-@withStyles(styles)
+@withStyles(s)
 class ContentPage extends Component {
 
   static propTypes = {
@@ -20,11 +20,9 @@ class ContentPage extends Component {
   render() {
     this.context.onSetTitle(this.props.title);
     return (
-      <div className="ContentPage">
-        <div className="ContentPage-container">
-          {
-            this.props.path === '/' ? null : <h1>{this.props.title}</h1>
-          }
+      <div className={s.root}>
+        <div className={s.container}>
+          {this.props.path === '/' ? null : <h1>{this.props.title}</h1>}
           <div dangerouslySetInnerHTML={{__html: this.props.content || ''}} />
         </div>
       </div>