+++ src/components/Footer/Footer.js
@@ -1,39 +1,24 @@
 /*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
 
-import React, { PropTypes, Component } from 'react';
-import styles from './Footer.scss';
-import withViewport from '../../decorators/withViewport';
+import React, { Component } from 'react';
+import s from './Footer.scss';
 import withStyles from '../../decorators/withStyles';
 import Link from '../Link';
 
-@withViewport
-@withStyles(styles)
+@withStyles(s)
 class Footer extends Component {
 
-  static propTypes = {
-    viewport: PropTypes.shape({
-      width: PropTypes.number.isRequired,
-      height: PropTypes.number.isRequired,
-    }).isRequired,
-  };
-
   render() {
-    // This is just an example how one can render CSS
-    const { width, height } = this.props.viewport;
-    this.renderCss(`.Footer-viewport:after {content:' ${width}x${height}';}`);
-
     return (
-      <div className="Footer">
-        <div className="Footer-container">
-          <span className="Footer-text">© Your Company</span>
-          <span className="Footer-spacer">·</span>
-          <a className="Footer-link" href="/" onClick={Link.handleClick}>Home</a>
-          <span className="Footer-spacer">·</span>
-          <a className="Footer-link" href="/privacy" onClick={Link.handleClick}>Privacy</a>
-          <span className="Footer-spacer">·</span>
-          <a className="Footer-link" href="/not-found" onClick={Link.handleClick}>Not Found</a>
-          <span className="Footer-spacer"> | </span>
-          <span ref="viewport" className="Footer-viewport Footer-text Footer-text--muted">Viewport:</span>
+      <div className={s.root}>
+        <div className={s.container}>
+          <span className={s.text}>© Your Company</span>
+          <span className={s.spacer}>·</span>
+          <a className={s.link} href="/" onClick={Link.handleClick}>Home</a>
+          <span className={s.spacer}>·</span>
+          <a className={s.link} href="/privacy" onClick={Link.handleClick}>Privacy</a>
+          <span className={s.spacer}>·</span>
+          <a className={s.link} href="/not-found" onClick={Link.handleClick}>Not Found</a>
         </div>
       </div>
     );