+++ src/components/Navigation/Navigation.js
@@ -1,12 +1,12 @@
 /*! React Starter Kit | MIT License | http://www.reactstarterkit.com/ */
 
-import React, { PropTypes, Component } from 'react';
-import classNames from 'classnames';
-import styles from './Navigation.scss';
+import React, { Component, PropTypes } from 'react';
+import cx from 'classnames';
+import s from './Navigation.scss';
 import withStyles from '../../decorators/withStyles';
 import Link from '../Link';
 
-@withStyles(styles)
+@withStyles(s)
 class Navigation extends Component {
 
   static propTypes = {
@@ -15,13 +15,13 @@ class Navigation extends Component {
 
   render() {
     return (
-      <div className={classNames(this.props.className, 'Navigation')} role="navigation">
-        <a className="Navigation-link" href="/about" onClick={Link.handleClick}>About</a>
-        <a className="Navigation-link" href="/contact" onClick={Link.handleClick}>Contact</a>
-        <span className="Navigation-spacer"> | </span>
-        <a className="Navigation-link" href="/login" onClick={Link.handleClick}>Log in</a>
-        <span className="Navigation-spacer">or</span>
-        <a className="Navigation-link Navigation-link--highlight" href="/register" onClick={Link.handleClick}>Sign up</a>
+      <div className={cx(s.root, this.props.className)} role="navigation">
+        <a className={s.link} href="/about" onClick={Link.handleClick}>About</a>
+        <a className={s.link} href="/contact" onClick={Link.handleClick}>Contact</a>
+        <span className={s.spacer}> | </span>
+        <a className={s.link} href="/login" onClick={Link.handleClick}>Log in</a>
+        <span className={s.spacer}>or</span>
+        <a className={cx(s.link, s.highlight)} href="/register" onClick={Link.handleClick}>Sign up</a>
       </div>
     );
   }