+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ### [Unreleased][unreleased]
 
+- Integrate [CSS Modules](https://github.com/css-modules/css-modules) and
+  [isomorphic-style-loader](https://github.com/kriasoft/isomorphic-style-loader)
 - Move `DOMUtils.js` to `src/core` folder; remove `src/utils` folder
 - Replace [cssnext](http://cssnext.io/) with [precss](https://github.com/jonathantneal/precss)
 - Update build automation scripts to use plain functions